-Better structure
